Add basic support for building for windows.
Add primitive support for building windows executables. This is not a
native port: a POSIX-like build environment is still assumed. With
MingW-W64, cross-compiling on linux or building on windows with MSYS2 or
Cygwin should be possible. In contrast to Cygwin builds, MingW-W64
executables are only dependent on `kernel32.dll` and `msvcrt.dll`, which
are included in the base system.

No accomodations were made to support the Visual C++ toolchain.